application/backend/freelancer-project/requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
Pillow==5.3.0
pytz==2018.5
